+++ bunny/__init__.py
@@ -1,2 +1,1 @@
 from .bunny import bunny
-from .bear import bear
